Add support for SHT30 temperature and humidity sensor
Lots of cleanup and reorganization of the file.
Added overall framework for storing multiple types of sensor readings
in the RTC memory.
Added readout of SHT30 temp/humidity sensor using Wemos library.
I previously had to modify this library but it is very simplistic so
I will likely remove the dependency and implemet the I2C transactions
directly.