Add support for Sharp GP2Y dust sensor
Detects the presence of the ~5k pullup on the LED output pin.
This requires an external 20k pulldown since the ESP8266 doesn't
really support internal pulldowns.

Performed some additional reorganization and increased the number
of stored sensor readings to 60.

The GP2Y library was at least available from the board manager
but is probably not worth keeping as a dependency long-term.
Will likely look for a better dust sensor anyway that has I2C.