Enable IOTAppStory Config Mode
Detects external resets outside of sleep mode (quick-ish pressing
the reset button after boot) and uses that as a trigger to enter
the configuration mode.

I also discovered that this was corrupting the ring buffer and
added the IAS RTC memory values to the enum so that we can avoid
interfering.