Remove IOTAppStory
Remove IOTAppStory usage with #ifdef IOTAPPSTORY.
This was proving to be more cumbersome that it was worth.
I was having to reimplement most of the features because it
didn't quite do what I wanted out-of-the-box and it was also
entering configuration mode way too easily and getting stuck
there.

Replaced it by adding a basic webserver in AP mode for to
handle future WiFi configuration and STA mode with reconnect
to upload measurements.

Took the opportunity to refactor some of the connectivity
functions.