Add unique build identifier
Define BUILD_UNIQUE_ID based on __TIME__ to serve as a unique
build identifier. It should always be < 16-bits.

Initial use for this is to detect when we upload new code so
that the RTC memory can be invalidated.
If the RTC memory is invalidated, use that as a sign to ignore
the Ext Reset press since this also happens every time we
upload new firmware and was often getting stuck in config mode.