Add OTA update capability
Added a firmware version identifier to the json preamble that
is sent to the report server with the sensor measurements.
This version identifier is based on minutes/seconds at build-
time. The report server will see if this matches the file that
it has in its firmware directory. If there is no match, it will
reply with an indication that a firmware update is available.

The filename should be formated as:
yyyymmdd-firmwareversion-filename.bin

Having received the update indication, the connectivity module
will send a new message (with the same json preamble)
requesting the firmware update be sent. The update package
consists of the file length (as a string) and the md5 checksum
(as a hex string) followed by the file contents.

After parsing the length and md5 checksum, the socket will be
passed to the ESP updateSketch API which handles the business
of the firmware update. This appears to be done in a robust
way -- I was not able to brick the device by pulling power
during the update process.