firmware update: invalidate rtc memory
If the firmware update begins in earnest the rtc user memory
needs to be invalidated. I was relying on the detection of new
firmware version to handle this on next boot, but if the update
fails it does not get invalidated resulting in the (possibly
corrupted) measurements being resent again immediately.