Implement configurable clock calibration
During deep sleep, there is significant drift of the clock due to
temperature which makes it difficult to determine the offset to add to
the uptime millis. This value is also likely to be different from one
module to another and, especially, between manufacturers of D1-mini
clones.

Over longer periods of time, the value can be averaged out and the
offset updated through the configuration portal.