Implement configurable sensor calibrations
These cheap sensors tend to have wide variance so that, while they are
precise, they are not always absolutely accurate. Additionally, they
will tend to change over time.
This becomes problematic when trying to compare several sensors against
each other or against past previous readings.

A calibration offset is thus applied to each sensor reading and this can
be modified in the configuration portal so that periodic calibrations
may be performed.
At the very least all sensors can be brought to a single location and
calibrated against each other. In other cases, they might be compared
against a known measurement standard.
Published NOAA readings should be suitable for pressure calibration.
A trusted mercury (or non-digital) thermometer for temperature.
And the well-known, if somewhat dubious, saturated salt water test for
humidity (in which the atmosphere of a closed container with a vial of
over-saturated salt water tends to equalize at about 75% humidity).