Add State-of-Health monitoring for node-red
Node-red seems to be quite unstable. Randomly (after a few
hours or a few days), it will lock-up and stop accepting data
from the sensor nodes. During this period the process shows
100% CPU usage, but no other diagnostic activity.

Add a flow that prints periodic messages to the log and a
monitor script that will watch for these messages and restart
node-red if they go absent for a few minutes.