Robustness: Add timeout to WiFi connection
The WiFi connection appears to take ~250 second to timeout in
some degenerate cases.
I'm guessing this is related to DHCP not getting a response,
but I wasn't able to get a good feel for this by changing
wifi_station_dhcpc_set_maxtry().

Instead, I have re-implemented waitForConnectResult() in order
to provide my own timeout of 30 seconds.