Robustness: implement v2 protocol
Implemented v2 of the TCP protocol for transmitting readings to
node-red. Incremented the version value in the json string and
created additional flows in node-red to separately parse the
new message format.

For robustness, it breaks up the json message into small batches
rather than sending all of the measurements together.
This also happens to better matche InfluxDB's measurement format
which wants to group time-correlated measurements together.

In order to pull batches of time-correlated measurements from the
RTC ring-buffer, there needs to be some flag at the end of the
batch. Additionally, only one timestamp is sent with each batch of
measurements so the per-measurement timestamp is now useless.
The new storage format removes the timestamp from each sensor
measurement and stores the uptime at the end of each batch.
The regular sensor measurements now fit in a single word and two
words are only needed for the uptime. Because of this, I treat the
uptime as 2 storage slots, one labeled SENSOR_TIMESTAMP_L and the
other SENSOR_TIMESTAMP_H representing the low and high 20-bits,
respectively.

Likewise, it is helpful for node-red to know when the final batch
of readings has been received so that it can group them all
together in a single InfluxDB transaction. I have appended the
uptime as an additional measurement in the final batch since it
seems like a useful thing to track but, anyway, is only meaningful
once per TCP session.

With the reduction of sensor readings from 2 slots to 1, we can
now store more than twice as many readings in the RTC buffer.
Likewise, the dump_readings function had to be updated to format
the data -- and, in particular, know how to combine low and high
timestamp bits into a single table row.