Tethered: fix uptime calculation when not sleeping
When we don't enter deep sleep, the millis value is not reset, but we
do keep using the stored value as the basis for our uptime calculation.
This leads to geometric growth in the uptime value as the millis value
is compounded with each call to save_rtc().

It is still useful to store the incremented uptime in RTC mem in case
of a warm reset. So, we can't just skip this step in tethered mode.

Instead save the old timestruct value and restore it after writing the
RTC user memory. If we get a warm reset, a reasonably up-to-date value
will exist in RTC memory. If we continue running without sleeping, the
original uptime value from the last reset will continue to be used for
uptime calculations.