Tethered: reset module if WiFi connection fails
It seems possible to get into a condition where WiFi can't connect and
a warm reboot is needed to fix it. My hypothesis is that something
changed in the AP config and the firmware doesn't know how to deal with
this and keeps trying to use the old values.

If the connection fails ask for a short deep_sleep to trigger a reboot.
Reconnect seems to work after this.