Battery: store timestamp using only 1 slot
Convert timestamp from using 2 24-bit sensor slots to just a single
slot by encoding it as an offset from a stored 32-bit base time.
The base time stores 256 ms ticks so that 40 bits of uptime can be
stored in a single 32-bit RTC slot.
When uploading readings to the server (or dumping them to the console)
the ms offset is added to the base ticks count to recreate the original
timestamp.