Battery: increase sleep time when upload fails
Increase the sleep time geometrically (up to 2^6 times) as long as
uploading the readings is failing to make progress.
In practice this should result in sleeping for as long as connectivity
has been unavailable -- up to 32 minutes. Power consumption should drop
equivalently and could approach 64x lower consumption for long outages.