node-red: Add config update feature and robustness
Allow remote update of config files in SPIFFS.
There are several robustness improvements here that were discovered
during the testing process. They are a little hard to separate out.

If there are files in sensor-cfg/<node-name>/, the server will reply
back to the device that it has config updates.
The device can then request the config files and update them in the
SPIFFS. If this process is successful, the device will request that
the config file be deleted from the server so that next time it
connects there will be no need to go through the process again.