connectivity: add config update feature
Handle new reply "config" reply from the server.
This will cause us to send a "get_config" command to the server to
request each of the known SPIFFS config files.
If the file is returned, it will have its md5sum checked and then be
written to SPIFFS. If this process is successful, a "delete_config"
command will be sent to the server to remove the now-stored config file.