connectivity: flush client buffer between commands
The old behavior of the WiFiClient seemed to mask the effect when some
partial data was left in the client receive buffer between transmitted
messages.
In particular, when requesting several calibration update files in a
row, the left over null byte from the previous message was being left
in the buffer resulting in the next request getting a null string as
the first response it saw.

Refactor the code to use a standard send_command() function that will
flush the Tx and Rx buffers, transmit the command, and verify the
transmit length is correct.