Fix issue with custom configs in WiFiManager
The custom configurations aren't saving in WiFiManager unless it gets an
updated SSID and is able to connect. In fact, it will not even exit the
configuration without a new SSID.

There is a new configuration option to restore the old behavior where it
exits the configurator any time the user saves the settings. However,
this will still typically fail to connect and return a failure.

Change to providing a callback function to be alerted when the custom
configs have been updated so that they can be stored regardless of the
connection/configuration status.