Add support for Waveshare 1.9" segmented E-Paper Display
Add EPD_1in9 library from Waveshare example code.
Add new disp_readings function to populate the display with (calibrated)
temperature and humidity readings.
There is also a BT icon that I am using to indicate when a connection to
the server is being attempted and a battery icon that I am setting when
the battery voltage drops below a configured threshold.