EPD_1in9: Only clear the display every 3 minutes
3 minute time was recommended in the data sheet.
However, this is probably an over-optimization.
The bulk of the time is spent waking the display from sleep and
this also seems to cause the screen to be cleared.

Still, this may save 2 seconds per boot when the temperature
is below 20 °C and maybe a bit more when it is below 15 °C.