EPD_1in9: Add 3D-model to display footprint
KiCad treats .wrl models as being based on .1" units for historical
reasons. Additionally, scaling and rotation in OpenSCAD result in loss of
color information in the .csg.

So, in the footprint properties, scaling for each axis is set to
0.393700787 and a 90 degree rotation is needed.
The translation of 49.35 mm brings it back in line with the KiCad origin.
(I suppose, scaling by a negative factor and rotation by -90° would also
do the trick.)