EPD_1in9: Create 3D Model for Display
Model is based on dimensions from https://www.waveshare.com/w/upload/5/53/1.9inch_Segment_e-Paper_Specification.pdf
with some measured dimensions and most segment are roughly approximated.

Workflow:
-Open the .scad in FreeCAD
-Wait a very long time...
-Select final 2 groups from the imported geometry
-Export as a .wrl file

(The OpenSCAD or CSG importer brings in all geometry and operations from
the .scad file and FreeCAD displays these all simultaneously. Visibility
for all of these except for the final 2 groups can be disabled to verify
that they get rendered properly.)

When you want to preview the model in OpenSCAD, change the FREECAD
variable at the top of the file to 0.